Best model for each criterion
Added option to save the best model for each criterion after epoch loop.
This is done by tracking the best epoch for each criterion.
At the end the model from the best epoch is copied to a file named
<ModelName>_<CriterionName>.
If training is interrupted best epoch per criterion log is saved in
checkpoint and reload on training resume.

Add E2E test for saveBestModelPerCriterion
Test created by combining Speech/DNN/ParallelCrossValidation and
Image/QuickE2E.
Test executes enough epochs so the last epoch is not always the best.
Test deletes the last epoch and starts the training again to simulate
a continuation of interrupted execution.
Generated baseline for Windows locally while for Linux baseline is from
Jenkins build with update baseline.